Move logic of createStoryFromTemplate to createStoryFromTemplateCallback
