+++ packages/wp-dashboard/package.json
@@ -15,6 +15,8 @@
         "@web-stories-wp/date": "*",
         "@web-stories-wp/design-system": "*",
         "@web-stories-wp/react": "*",
+	    "@web-stories-wp/patterns": "*",
+        "@web-stories-wp/story-editor": "*",
         "@web-stories-wp/tracking": "*",
         "@wordpress/api-fetch": "^5.2.1",
         "flagged": "^2.0.1"
