+++ packages/wp-story-editor/package.json
@@ -16,6 +16,7 @@
         "@web-stories-wp/i18n": "*",
         "@web-stories-wp/media": "*",
         "@web-stories-wp/migration": "*",
+        "@web-stories-wp/patterns": "*",
         "@web-stories-wp/react": "*",
         "@web-stories-wp/story-editor": "*",
         "@web-stories-wp/tracking": "*",
